use template dbs for faster schema creation

+++ packages/elefast-core/src/elefast_core/asyncio/metadata.py
@@ -15,6 +15,7 @@ class MetaDataBasedAsyncDatabaseServer(AsyncDatabaseServer):
     def __init__(self, engine: CanBeTurnedIntoAsyncEngine, metadata: MetaData) -> None:
         self._metadata = metadata
         self._engine = build_engine(engine)
+        self._template_db_name: str | None = None
 
     @override
     @property
@@ -23,10 +24,18 @@ class MetaDataBasedAsyncDatabaseServer(AsyncDatabaseServer):
 
     @override
     async def create_database(self) -> AsyncDatabase:
-        engine = await prepare_async_database(self._engine)
-        async with engine.begin() as connection:
-            await connection.run_sync(self._metadata.drop_all)
-            await connection.run_sync(self._metadata.create_all)
+        template_db = self._template_db_name
+        if template_db is None:
+            engine = await prepare_async_database(self._engine)
+            async with engine.begin() as connection:
+                await connection.run_sync(self._metadata.drop_all)
+                await connection.run_sync(self._metadata.create_all)
+            await engine.dispose()
+            template_db = engine.url.database
+            assert isinstance(template_db, str)
+            self._template_db_name = template_db
+
+        engine = await prepare_async_database(self._engine, template=template_db)
         return AsyncDatabase(engine=engine, server=self)
 
     @override
